[Ide] Properly handle symlinks at workbench level
When working with symlinked on disk files in a project, there's a
few things which were not handled correctly.

When a project references a symlink, we should keep the original
symlink information around so that it can be used when looking up
a for a Project containing a file or a file in general in a project.

Added in an OriginalFile property that is used to recover the original
file name, before resolving a symlink, that can be used by the typesystem
to figure out which project the file belongs to.

Fixes VSTS #700204 - Symlinked files do not have correct define constants